requirements.txt
PyYAML==3.13
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Pillow==8.4.0
tensorflow-gpu==1.14.0
h5py==2.8.0
pandas==1.1.5